LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-grid-flex-multicol-expected.txt

12

1
2

1
2

1
2

1
2

1
2

PASS flex
FAIL inline-flex assert_equals: display expected "inline-flex" but got "flex"
PASS grid
FAIL inline-grid assert_equals: display expected "inline-grid" but got "grid"
PASS multicol
